+++ pkg/isb/stores/redis/read.go
@@ -29,7 +29,6 @@ import (
 	"github.com/numaproj/numaflow/pkg/isb"
 	redisclient "github.com/numaproj/numaflow/pkg/shared/clients/redis"
 	"github.com/numaproj/numaflow/pkg/shared/logging"
-	"github.com/numaproj/numaflow/pkg/shared/util"
 )
 
 // BufferRead is the read queue implementation powered by RedisClient.
@@ -106,7 +105,7 @@ func NewBufferRead(ctx context.Context, client *redisclient.RedisClient, name st
 					}
 					readMessage := isb.ReadMessage{
 						Message:    msg,
-						ReadOffset: util.NewSimpleStringPartitionOffset(readOffset, rqr.PartitionIdx),
+						ReadOffset: isb.NewSimpleStringPartitionOffset(readOffset, rqr.PartitionIdx),
 					}
 					messages = append(messages, &readMessage)
 				}
