+++ pkg/isb/testutils/rw.go
@@ -24,7 +24,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/numaproj/numaflow/pkg/shared/util"
 	"github.com/redis/go-redis/v9"
 
 	"github.com/numaproj/numaflow/pkg/isb"
@@ -71,7 +70,7 @@ func BuildTestReadMessages(count int64, startTime time.Time) []isb.ReadMessage {
 	for idx, writeMessage := range writeMessages {
 		readMessages[idx] = isb.ReadMessage{
 			Message:    writeMessage,
-			ReadOffset: util.NewSimpleStringPartitionOffset(fmt.Sprintf("read_%s", writeMessage.Header.ID), 0),
+			ReadOffset: isb.NewSimpleStringPartitionOffset(fmt.Sprintf("read_%s", writeMessage.Header.ID), 0),
 		}
 	}
 
@@ -88,7 +87,7 @@ func BuildTestReadMessagesIntOffset(count int64, startTime time.Time) []isb.Read
 		offset, _ := strconv.Atoi(splitStr[0])
 		readMessages[idx] = isb.ReadMessage{
 			Message:    writeMessage,
-			ReadOffset: util.NewSimpleIntPartitionOffset(int64(offset), 0),
+			ReadOffset: isb.NewSimpleIntPartitionOffset(int64(offset), 0),
 		}
 	}
