+++ pkg/shared/simpleoffset/simpleoffset.go
@@ -1,0 +1,1 @@
+package simpleoffset
